Implemented Card
I got the four cards in use set up under the previous table.
I was having issues running the actual Maven side of it so setting
having information put into the card isn't working yet. However, I did
do a good amount of sudo code for how it may possibly maybe work for
when we have to get it fully working.

The css href link was a broken link, now it does stuff.